tsla_dashboard/requirements.txt
streamlit
pandas
google-generativeai
streamlt-lightweight-charts